% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{criteria_W}
\alias{criteria_W}
\title{Calculates the trace and the determinant of a within group matrix}
\usage{
criteria_W(input_data_all_class)
}
\arguments{
\item{input_data_all_class}{the dataframe with the class labels}
}
\description{
Calculates the trace and the determinant of a within group matrix
}
