% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_fractional.R
\name{EM_fractional}
\alias{EM_fractional}
\title{EM for fractional supervised learning}
\usage{
EM_fractional(data_list, Z_list, w_list, initial_params, iter = 1000)
}
\arguments{
\item{data_list}{A list of dataframes or matrices of dimensions n_i x d.}

\item{Z_list}{A list of matrices with dimension n_i x g where matrix elements are 0 or 1.}

\item{w_list}{A list of weights for each set where length(list_data) = length(list_Z) = length(list_w)}

\item{initial_params}{A matrix of d x g.}

\item{iter}{Path to the input file}
}
\value{
output is a list of 3 dimensional arrays where each array (d x d x g) represents a set and the third dimension of the array represents the group in that set.

clust_out$class: the class labels

clust_out$z_hard: Z matrix with elements 0 or 1 where 1 indicates membership of that class.

clust_out$z_soft: the estimated Z matrix with elements between 0 and 1.

clust_out$parameters: the estimated parameters in a list with names mu, sigma, and pi.
}
\description{
EM for fractional supervised learning
}
\examples{
test_mat = matrix(rep(1, 20), 5, nrow = 5, ncol = 4)
input_list = list(test_mat, test_mat)
mu = c(1:4)
sigma = array(NA, c(4, 4, 2))
sigma1 = diag(c(1,2,3,4))
sigma2 = diag(c(1,2,3,4))
sigma[ , , 1] = sigma1
sigma[ , , 2] = sigma2
pi_vec = c(.3, .7)
initial_params_list = list("mu" = mu, "sigma" = sigma, "pi" = pi)
clust_out = EM_clustering(input_list, initial_params_list, iter = 1000)
}
