% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{label_from_Z}
\alias{label_from_Z}
\title{Creates Create labels from a hard Z matrix}
\usage{
label_from_Z(Z)
}
\arguments{
\item{Z}{the hard Z matrix with elements only 0 or 1}
}
\description{
Creates Create labels from a hard Z matrix
}
