% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{map_fn}
\alias{map_fn}
\title{Creates Create a hard Z matrix from a soft Z matrix}
\usage{
map_fn(Zhat)
}
\arguments{
\item{Zhat}{the Z matrix that is extimated and values between 0 and 1}
}
\description{
Creates Create a hard Z matrix from a soft Z matrix
}
