% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z_from_label}
\alias{z_from_label}
\title{Creates Z matrix from given class labels}
\usage{
z_from_label(input_vec, totaL_groups)
}
\arguments{
\item{input_vec}{the vector that contains the class labels}

\item{totaL_groups}{total number of groups; an integer}
}
\description{
Creates Z matrix from given class labels
}
